+++ packages/@expo/cli/e2e/fixtures/with-assets/app.config.js
@@ -10,7 +10,4 @@ module.exports = {
   web: {
     bundler: 'metro',
   },
-  experiments: {
-    tsconfigPaths: process.env._EXPO_E2E_USE_PATH_ALIASES ? true : undefined,
-  },
 };