+++ packages/@expo/cli/e2e/fixtures/with-router/app.config.js
@@ -12,7 +12,6 @@ module.exports = {
     bundler: 'metro',
   },
   experiments: {
-    tsconfigPaths: process.env._EXPO_E2E_USE_PATH_ALIASES ? true : undefined,
-    typedRoutes: process.env._EXPO_E2E_USE_TYPED_ROUTES ? true : undefined
+    typedRoutes: process.env._EXPO_E2E_USE_TYPED_ROUTES ? true : undefined,
   },
 };