+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -105,7 +105,6 @@ export async function loadMetroConfigAsync(
     routerDirectory: getRouterDirectoryModuleIdWithManifest(projectRoot, exp),
     config,
     platformBundlers,
-    isTsconfigPathsEnabled: !!exp.experiments?.tsconfigPaths,
     webOutput: exp.web?.output ?? 'single',
   });