+++ packages/@expo/cli/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### 🛠 Breaking changes
 
+- Enable `tsconfigPaths` and remove `expo.experiments.tsconfigPaths` flag
+
 ### 🎉 New features
 
 ### 🐛 Bug fixes